cleanup: consolidate requestdataproducer into dataproducer
The requestcontrol plugin tree had two sibling folders serving the same
role: dataproducer/ (approximateprefix) and requestdataproducer/
(inflightload, predictedlatency). Move inflightload and predictedlatency
under dataproducer/ and drop the now-empty requestdataproducer/ directory.
Import paths and the framework-imports exception list are updated to match.
